feat(COURT) Add authentication to some disable-courts routes

+++ src/courts/disable-courts/disable-courts.controller.ts
@@ -1,4 +1,5 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query } from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpException, HttpStatus, Param, Post, Put, Query, Req, UseGuards } from '@nestjs/common';
+import { JwtAuthGuard } from 'src/auth/jwt.guard';
 import { CreateDisableCourtDTO } from './disable-courts.dto';
 import { DisableCourtsService } from './disable-courts.service';
 import { DisableCourt } from './interfaces/disable-courts.interface';
@@ -6,33 +7,40 @@ import { DisableCourt } from './interfaces/disable-courts.interface';
 @Controller('courts/disable-courts')
 export class DisableCourtsController {
     constructor(private readonly disableCourtsService: DisableCourtsService) { }
+
+    @UseGuards(JwtAuthGuard)
     @Post('')
-    async createDisableCourt(@Body() body: CreateDisableCourtDTO, @Query('merge') merge: boolean): Promise<DisableCourt> {
+    async createDisableCourt(@Req() req, @Body() body: CreateDisableCourtDTO, @Query('merge') merge: boolean): Promise<DisableCourt> {
+        if (!req.user.isStaff) throw new HttpException("Not a Staff", HttpStatus.UNAUTHORIZED);
         return await this.disableCourtsService.createDisableCourt(body, merge);
     }
 
     @Get('')
-    async getAllDisableCourt(): Promise<DisableCourt[]> {
+    async getAllDisableCourt(@Req() req): Promise<DisableCourt[]> {
         return await this.disableCourtsService.getAllDisableCourt();
     }
-    
+
     @Get('closed_time')
-    async getClosedTime(@Query('sport_name') sport_name: string, @Query('court_num') court_num: number, @Query('date') dateString: string): Promise<Array<[number,number]>> {
-        return await this.disableCourtsService.findClosedTimes(sport_name,court_num,new Date(dateString));
+    async getClosedTime(@Req() req, @Query('sport_name') sport_name: string, @Query('court_num') court_num: number, @Query('date') dateString: string): Promise<Array<[number, number]>> {
+        return await this.disableCourtsService.findClosedTimes(sport_name, court_num, new Date(dateString));
     }
 
     @Get(':id')
-    async getDisableCourt(@Param('id') id: string): Promise<DisableCourt> {
+    async getDisableCourt(@Req() req, @Param('id') id: string): Promise<DisableCourt> {
         return await this.disableCourtsService.getDisableCourt(id);
     }
 
+    @UseGuards(JwtAuthGuard)
     @Delete('')
-    async deleteAllDisableCourt(): Promise<void> {
+    async deleteAllDisableCourt(@Req() req): Promise<void> {
+        if (!req.user.isStaff) throw new HttpException("Not a Staff", HttpStatus.UNAUTHORIZED);
         await this.disableCourtsService.deleteAllDisableCourt();
     }
 
+    @UseGuards(JwtAuthGuard)
     @Delete(':id')
-    async deleteDisableCourt(@Param('id') id: string): Promise<void> {
+    async deleteDisableCourt(@Req() req, @Param('id') id: string): Promise<void> {
+        if (!req.user.isStaff) throw new HttpException("Not a Staff", HttpStatus.UNAUTHORIZED);
         await this.disableCourtsService.deleteDisableCourt(id);
     }
